rename analytics constanst file to constants and change email and letter strings to constants
