+++ lib/use-cases/create-conversation.js
@@ -1,4 +1,5 @@
 import { convertObjectToIsoDate } from 'lib/utils/date';
+import { EMAIL, LETTER } from 'lib/utils/constants';
 
 export default class CreateConversation {
   constructor({ conversationsGateway, sendSignpostEmail, sendSignpostLetter }) {
@@ -16,7 +17,7 @@ export default class CreateConversation {
     const errors = [];
 
     let signPostingEmail;
-    if (sharingMethod == 'email') {
+    if (sharingMethod == EMAIL) {
       signPostingEmail = await this.sendSignpostEmail.execute({
         emailAddress: resident.email,
         signPostingMessage
@@ -26,7 +27,7 @@ export default class CreateConversation {
       }
     }
     let signPostingLetter;
-    if (sharingMethod == 'letter') {
+    if (sharingMethod == LETTER) {
       signPostingLetter = await this.sendSignpostLetter.execute({
         name: `${resident.firstName} ${resident.lastName}`,
         address: resident.address,
