+++ lib/use-cases/create-conversation.test.js
@@ -1,4 +1,5 @@
 import CreateConversation from 'lib/use-cases/create-conversation';
+import { EMAIL, LETTER } from 'lib/utils/constants';
 
 describe('Create Conversation use case', () => {
   let sendResidentSms;
@@ -91,7 +92,7 @@ describe('Create Conversation use case', () => {
         discussedServices: discussedServices,
         signPostingMessage: signPostingMessage
       },
-      'email'
+      EMAIL
     );
 
     expect(sendSignpostEmail.execute).toHaveBeenCalledWith({
@@ -151,7 +152,7 @@ describe('Create Conversation use case', () => {
         discussedServices: discussedServices,
         signPostingMessage: signPostingMessage
       },
-      'letter'
+      LETTER
     );
 
     expect(sendSignpostEmail.execute).not.toHaveBeenCalled();
