+++ lib/utils/analytics.js
@@ -1,5 +1,5 @@
 /* eslint-disable no-undef */
-import { ANALYTICS_GROUPS } from 'lib/utils/analyticsConstants';
+import { ANALYTICS_GROUPS } from 'lib/utils/constants';
 
 const sendDataToAnalytics = ({
   action,
