+++ lib/utils/constants.js
@@ -19,3 +19,7 @@ export const ANALYTICS_GROUPS = {
   'Better Conversations User Group 5': 'Early Help',
   Other: 'Other'
 };
+
+export const EMAIL = 'email';
+
+export const LETTER = 'letter';
