+++ lib/utils/getEmailBody.js
@@ -1,3 +1,5 @@
+import { LETTER } from 'lib/utils/constants';
+
 const getEmailBody = (residentInfo, newSignpostSummary, newReferralSummary, referrerData, type) => {
   const discussed =
     newSignpostSummary.length > 0
@@ -15,7 +17,7 @@ const getEmailBody = (residentInfo, newSignpostSummary, newReferralSummary, refe
           .join('')}`
       : '';
 
-  if (type == 'letter') {
+  if (type == LETTER) {
     return `Hi ${residentInfo.firstName || ''} ${residentInfo.lastName || ''},
 
 It was nice speaking with you. I am writing to share the contact details of the support services we discussed in our conversation.
